refactor: move user entity to domain
